requirements.txt
# Core RAG
langchain
langchain-groq
langchain-huggingface
langchain-community
faiss-cpu
pypdf
sentence-transformers
torch
torchvision

# API & Environment
python-dotenv
groq
google-genai

# UI
streamlit
watchdog
